fixed db issue
