+++ .gitignore
@@ -169,6 +169,4 @@ cython_debug/
 
 # PyPI configuration file
 .pypirc
-/auth_api/migrations/
-/subjects/migrations/
-/generic_test/migrations/
+
