+++ requirements.txt
@@ -1,6 +1,5 @@
 annotated-types==0.7.0
 asgiref==3.8.1
-async-timeout==5.0.1
 black==25.1.0
 certifi==2025.1.31
 cffi==1.17.1
