+++ subject/apps.py
@@ -3,4 +3,4 @@ from django.apps import AppConfig
 
 class SubjectsConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
-    name = "subjects"
+    name = "subject"
