+++ subject/urls.py
@@ -1,8 +1,8 @@
 from django.urls import path
 
-from subjects.views.create_subjects import CreateSubjectView
-from subjects.views.edit_subject import EditSubjectView
-from subjects.views.get_all_subjects import AllSubjectsView
+from subject.views.create_subjects import CreateSubjectView
+from subject.views.edit_subject import EditSubjectView
+from subject.views.get_all_subjects import AllSubjectsView
 
 urlpatterns = [
     path("create-subject", CreateSubjectView.as_view(), name="Create-subject"),
